+++ src/components/NewGame/NewGameDialog/index.tsx
@@ -9,7 +9,7 @@ import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
 
 import { GameDifficulty, GameMode, WordLength } from "components/Game/types";
-import { useCreateNewGame } from "core/game";
+import { useCreateNewGame } from "components/NewGame/utils";
 import ModeSettings from "components/NewGame/NewGameDialog/ModeSettings";
 import LettersSettings from "components/NewGame/NewGameDialog/LettersSettings";
 import NumbersSettings from "components/NewGame/NewGameDialog/NumbersSettings";
